cast signal number

+++ main.c
@@ -103,12 +103,12 @@ int main(int argc, char *argv[]) {
             fprintf(stderr, "BRIGHT environment variable not set.\n");
             return 0;
         }
-        if ((BRIGHT.number = atol(BRIGHT.string)) < 10) {
+        if ((BRIGHT.number = (uint64) atol(BRIGHT.string)) < 10) {
             fprintf(stderr, "Invalid BRIGHT environment variable: %s.\n", BRIGHT.string);
             return 0;
         }
 
-        send_signal(program_to_signal, BRIGHT.number);
+        send_signal(program_to_signal, (int) BRIGHT.number);
     }
 
     return 0;
